fix: Partially resolves #3: Avoid sending diff content to the editor when it remains unchanged
It seems that editor.execCommand focuses the Rich Text Editor when
called. It also seems to be the way to send content to the Markdown
editor. This commit causes the plugin to use editor.execCommand less
often, thereby causing focus to jump to the Rich Text Editor less often.
This is not a complete fix, which may require changes to Joplin.